COM-337 COM-265 changes

See merge request team-monitoring/comedian!12

+++ chat/slack.go
@@ -7,11 +7,11 @@ import (
 	"time"
 
 	"github.com/jasonlvhit/gocron"
-	"github.com/maddevsio/comedian/config"
-	"github.com/maddevsio/comedian/model"
-	"github.com/maddevsio/comedian/storage"
 	"github.com/nlopes/slack"
 	"github.com/sirupsen/logrus"
+	"gitlab.com/team-monitoring/comedian/config"
+	"gitlab.com/team-monitoring/comedian/model"
+	"gitlab.com/team-monitoring/comedian/storage"
 
 	"strings"
 )
@@ -53,7 +53,7 @@ func (s *Slack) Run() {
 	s.UpdateUsersList()
 	s.SendUserMessage(s.Conf.ManagerSlackUserID, s.Conf.Translate.HelloManager)
 
-	gocron.Every(1).Day().At("23:50").Do(s.FillStandupsForNonReporters)
+	gocron.Every(1).Day().At("23:59").Do(s.FillStandupsForNonReporters)
 	gocron.Every(1).Day().At("23:55").Do(s.UpdateUsersList)
 	gocron.Start()
 
@@ -128,9 +128,7 @@ func (s *Slack) handleMessage(msg *slack.MessageEvent, botUserID string) {
 			}
 			logrus.Infof("Standup created #id:%v\n", standup.ID)
 			item := slack.ItemRef{msg.Channel, msg.Msg.Timestamp, "", ""}
-			time.Sleep(2 * time.Second)
 			s.API.AddReaction("heavy_check_mark", item)
-			s.SendEphemeralMessage(msg.Channel, msg.User, s.Conf.Translate.StandupHandleCreatedStandup)
 			return
 		}
 	case typeEditMessage:
@@ -165,9 +163,8 @@ func (s *Slack) handleMessage(msg *slack.MessageEvent, botUserID string) {
 				}
 				logrus.Infof("Standup created #id:%v\n", standup.ID)
 				item := slack.ItemRef{msg.Channel, msg.SubMessage.Timestamp, "", ""}
-				time.Sleep(2 * time.Second)
+
 				s.API.AddReaction("heavy_check_mark", item)
-				s.SendEphemeralMessage(msg.Channel, msg.SubMessage.User, s.Conf.Translate.StandupHandleCreatedStandup)
 				return
 			}
 		}
@@ -179,10 +176,13 @@ func (s *Slack) handleMessage(msg *slack.MessageEvent, botUserID string) {
 		}
 		if messageIsStandup {
 			standup.Comment = msg.SubMessage.Text
-			st, _ := s.DB.UpdateStandup(standup)
+			st, err := s.DB.UpdateStandup(standup)
+			if err != nil {
+				logrus.Errorf("UpdateStandup failed: %v", err)
+				s.SendEphemeralMessage(msg.Channel, msg.SubMessage.User, s.Conf.Translate.StandupHandleCouldNotSaveStandup)
+				return
+			}
 			logrus.Infof("Standup updated #id:%v\n", st.ID)
-			time.Sleep(2 * time.Second)
-			s.SendEphemeralMessage(msg.Channel, msg.SubMessage.User, s.Conf.Translate.StandupHandleUpdatedStandup)
 			return
 		}
 
@@ -191,8 +191,12 @@ func (s *Slack) handleMessage(msg *slack.MessageEvent, botUserID string) {
 		if err != nil {
 			logrus.Errorf("SelectStandupByMessageTS failed: %v", err)
 		}
-		s.DB.DeleteStandup(standup.ID)
-		logrus.Infof("Standup deleted #id:%v\n", standup.ID)
+		if standup.UserID == msg.User {
+			s.DB.DeleteStandup(standup.ID)
+			logrus.Infof("Standup deleted #id:%v\n", standup.ID)
+		} else {
+			logrus.Infof("Could not delete standup #id:%v. User is not the owner!\n", standup.ID)
+		}
 	}
 }
 
@@ -322,19 +326,21 @@ func (s *Slack) UpdateUsersList() {
 //FillStandupsForNonReporters fills standup entries with empty standups to later recognize
 //non reporters vs those who did not have to write standups
 func (s *Slack) FillStandupsForNonReporters() {
+	logrus.Info("FillStandupsForNonReporters begin")
 	if int(time.Now().Weekday()) == 6 || int(time.Now().Weekday()) == 0 {
+		logrus.Info("Weekends! Do not check!")
 		return
 	}
 	allUsers, err := s.DB.ListAllChannelMembers()
 	if err != nil {
+		logrus.Errorf("ListAllChannelMembers while FillStandupsForNonReporters failed: %v", err)
 		return
 	}
 	for _, user := range allUsers {
-		if user.Created.Day() == time.Now().Day() {
-			continue
-		}
 		hasStandup := s.DB.SubmittedStandupToday(user.UserID, user.ChannelID)
-		if !hasStandup {
+		shouldBeTracked := s.DB.MemberShouldBeTracked(user.ID, time.Now())
+		logrus.Infof("User [%v] in [%v] should be tracked [%v] and has standup [%v]", user.UserID, user.ChannelID, shouldBeTracked, hasStandup)
+		if !hasStandup && shouldBeTracked {
 			_, err := s.DB.CreateStandup(model.Standup{
 				ChannelID: user.ChannelID,
 				UserID:    user.UserID,
@@ -345,6 +351,7 @@ func (s *Slack) FillStandupsForNonReporters() {
 				errorReportToManager := fmt.Sprintf("I could not create empty standup for user %s in channel %s because of the following reasons: %v", user.UserID, user.ChannelID, err)
 				s.SendUserMessage(s.Conf.ManagerSlackUserID, errorReportToManager)
 			}
+			logrus.Infof("Empty standup created for user [%v] in [%v]", user.UserID, user.ChannelID)
 		}
 	}
 }
